Rename paths, use commonjs2 as build target
